SM-99: rename project-contents to project
