+++ bitwarden_license/bit-web/src/app/secrets-manager/projects/project/project-secrets.component.html
