+++ bitwarden_license/bit-web/src/app/secrets-manager/projects/project/project.component.html
