+++ bitwarden_license/bit-web/src/app/secrets-manager/projects/projects.module.ts
@@ -4,8 +4,8 @@ import { SecretsManagerSharedModule } from "../shared/sm-shared.module";
 
 import { ProjectDeleteDialogComponent } from "./dialog/project-delete-dialog.component";
 import { ProjectDialogComponent } from "./dialog/project-dialog.component";
-import { ProjectContentsComponent } from "./project-contents/project-contents.component";
-import { ProjectSecretsComponent } from "./project-contents/project-secrets.component";
+import { ProjectSecretsComponent } from "./project/project-secrets.component";
+import { ProjectComponent } from "./project/project.component";
 import { ProjectsListComponent } from "./projects-list/projects-list.component";
 import { ProjectsRoutingModule } from "./projects-routing.module";
 import { ProjectsComponent } from "./projects/projects.component";
@@ -17,7 +17,7 @@ import { ProjectsComponent } from "./projects/projects.component";
     ProjectsListComponent,
     ProjectDialogComponent,
     ProjectDeleteDialogComponent,
-    ProjectContentsComponent,
+    ProjectComponent,
     ProjectSecretsComponent,
   ],
   providers: [],
